chore: :truck: moving factory structures to the domain
